+++ source/state_representation/src/space/cartesian/CartesianState.cpp
@@ -538,7 +538,9 @@ CartesianState CartesianState::inverse() const {
   inverse.set_angular_velocity(inverse_angular_velocity);
   inverse.set_linear_acceleration(inverse_linear_acceleration);
   inverse.set_angular_acceleration(inverse_angular_acceleration);
-  // TODO(#30): wrench inverse
+
+  // the inverse wrench is not supported by this operation
+  inverse.set_wrench(Eigen::Vector<double, 6>::Zero());
 
   return inverse;
 }
@@ -639,6 +641,7 @@ CartesianState& CartesianState::operator*=(const CartesianState& state) {
     throw IncompatibleReferenceFramesException("Expected " + this->get_name() + ", got " + state.get_reference_frame());
   }
   this->set_name(state.get_name());
+
   // intermediate variables for f_S_b
   Eigen::Vector3d f_P_b = this->get_position();
   Eigen::Quaterniond f_R_b = this->get_orientation();
@@ -646,31 +649,40 @@ CartesianState& CartesianState::operator*=(const CartesianState& state) {
   Eigen::Vector3d f_omega_b = this->get_angular_velocity();
   Eigen::Vector3d f_a_b = this->get_linear_acceleration();
   Eigen::Vector3d f_alpha_b = this->get_angular_acceleration();
+
   // intermediate variables for b_S_c
   Eigen::Vector3d b_P_c = state.get_position();
-  // specific operation on quaternion using Hamilton product, keeping the resulting quaternion on the same hemisphere
   Eigen::Quaterniond b_R_c = state.get_orientation();
   Eigen::Vector3d b_v_c = state.get_linear_velocity();
   Eigen::Vector3d b_omega_c = state.get_angular_velocity();
   Eigen::Vector3d b_a_c = state.get_linear_acceleration();
   Eigen::Vector3d b_alpha_c = state.get_angular_acceleration();
+  Eigen::Vector3d b_F_c = state.get_force();
+  Eigen::Vector3d b_tau_c = state.get_torque();
   // pose
   this->set_position(f_P_b + f_R_b * b_P_c);
   auto orientation = f_R_b * b_R_c;
+
+  // specific operation on quaternion using Hamilton product, keeping the resulting quaternion on the same hemisphere
   if (orientation.dot(this->get_orientation()) < 0) {
     orientation = Eigen::Quaterniond(-orientation.coeffs());
   }
   this->set_orientation(orientation);
+
   // twist
   this->set_linear_velocity(f_v_b + f_R_b * b_v_c + f_omega_b.cross(f_R_b * b_P_c));
   this->set_angular_velocity(f_omega_b + f_R_b * b_omega_c);
+
   // acceleration
   this->set_linear_acceleration(
       f_a_b + f_R_b * b_a_c + f_alpha_b.cross(f_R_b * b_P_c) + 2 * f_omega_b.cross(f_R_b * b_v_c)
           + f_omega_b.cross(f_omega_b.cross(f_R_b * b_P_c)));
   this->set_angular_acceleration(f_alpha_b + f_R_b * b_alpha_c + f_omega_b.cross(f_R_b * b_omega_c));
-  // wrench
-  //TODO
+
+  // keep only the wrench measured at the distal frame, aligned with the new reference frame
+  this->set_force(f_R_b * b_F_c);
+  this->set_torque(f_R_b * b_tau_c);
+
   return (*this);
 }